+++ ui/step_editor/http_request_config_dialog.py
@@ -20,6 +20,7 @@ class _HttpRequestConfigDialog(_BaseStepConfigDialog):
     def __init__(self, config: dict, parent=None, label: str = "", **_):
         super().__init__(config, parent, label,
                           retry_count=_.get("retry_count", 0),
+                          retry_interval_s=_.get("retry_interval_s", 5),
                           run_always=_.get("run_always", False),
                           timeout_s=_.get("timeout_s", 0))
         self._prior_steps = _.get("prior_steps") or []